Another attempt at loose coupling
I decided to not roll the system back after all, and fixed all of the
issues I encountered in the last commit.
I have not tested whether the system still works if there are more than
one type of Node/Way available, but at least everything works with all
of the interfaces I have created. Note that there are many type-casts in
my code - some of which may not be safe (but I haven't tested that
either)...