More of the same
Forgot to merge the last version...
Nodes and Ways are now stored in a static array in BuildDatabase, and
indexed with node.getLocalId() or way.getLocalId().
All tests are completing successfully now.
Some issues with infinite loops in AStar.getPath() where nodes point to
themselves or back at their parent.
AStarTest has an issue where the path found is correct in one test, but
missing a step in another... Marked with TODO and comments explaining
further.