Broken AStar
Fixed some aspects of the map, including the ability to click on it to
recalculate a route from where you clicked.
A* seems to be broken after I implemented an isTower method in Node, and
changed AStar.class' code to use it... The routes it generates just
suddenly jumps; This problem did not occur until I implemented
click-functionality on the map, and the isTower method in Node.