Fixed navigation, storage, and filtering
Every algorithm works as it should now. Only Nodes suitable for
navigation are stored after filtering(i.e. no stairs, non-automatic
doors, etc.). The task of filtering has been moved to Main- be wary of
this. BuildDatabase does not delete any Nodes or Ways - besides
unconnected, isolated ones, and inaccessible doors (it is probably
possible to filter inaccessible doors elsewhere though).
Persisting issues:
Map does not display route-distance
NullPointerException (and a few other exceptions) thrown by Mapsforge's
api; Unrelated to my implementation (as far as I know).
Only four algorithms implemented.
GBFS is not complete - does not have to be, but mention that this is
intentional in the report.
No option to select disability; filtering made for (manual?) wheelchairs
only.
Comments and console-output still present in the code. These should be
removed before the final version is handed in.